pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-mvp"
version = "0.1.0"
description = "Production-credible MVP for a multi-agent automated trading platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "python-dotenv>=1.0.1",
  "psycopg[binary]>=3.2.3",
  "redis>=5.2.0",
  "rq>=1.16.2",
  "httpx>=0.28.1",
  "cryptography>=44.0.2",
]

[project.optional-dependencies]
dev = [
  "beautifulsoup4>=4.12.3",
  "markdown>=3.7",
  "pillow>=11.1.0",
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=0.24.0",
  "reportlab>=4.2.5",
  "ruff>=0.8.0",
  "mypy>=1.13.0",
]

[tool.setuptools.packages.find]
where = ["backend"]
include = ["trading_mvp*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008", "E501", "SIM102"]

[tool.mypy]
python_version = "3.12"
mypy_path = "backend"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
strict_equality = true
plugins = []
